refactor: UserService 수정
- UserService 인터페이스 삭제
- UserServiceImpl -> UserService 변경
- validateExistedUser() 생성 : 중복 유저 확인 메서드
- 프로덕션 코드 파라미터에 final 키워드 추가
