+++ src/main/java/com/flab/just_10_minutes/User/domain/User.java
@@ -15,7 +15,7 @@ public class User {
     private String phone;
     private String address;
     private ROLE role;
-    //TODO : point 추가
+    //TODO : point 필드 추가 -> Point 관리 로직 생성 후 적용 예정(7/14)
 
     public enum ROLE {
         PUBLIC,
